fix: enable host networking and root users on dev
