+++ packages/services/ds/src/ops/game_config/mod.rs
@@ -1,2 +1,2 @@
 pub mod get;
-pub mod update;
+pub mod upsert;
